Preserve trailing slash in URL path
The previous behavior unconditionally stripped a trailing slash from all
URLs, which resulted in parsed URLs differing from the original:

    http://example.com/path/ -> http://example.com/path

Now the original URL is preserved. Furthermore, these two URLs are no
longer considered to be equal by the matches() method.

However, a trailing slash is still stripped from an "empty" path, which
is in keeping with the original behavior:

    http://example.com/ -> http://example.com